Added link to contact admins, changed wording, and moved error redirects to app/__init__.py
