+++ app/templates/errors/403error.html
@@ -6,7 +6,8 @@
 <div class="text-center">
     <h1 class="display-1">Error 403</h1>
 
-    <p class="">You do not have permissions to view this page. If you believe this is a mistake, please contact the system administrator.</p>
+    <p class="">You do not have permissions to view this page. If you believe this is a mistake,
+                <a href="mailto:heggens@berea.edu,ramsayb2@berea.edu">please contact the system administrators.</a></p>
     <br>
     <a class="btn btn-primary" href="/" role=button>Return Home</a>
 </div>
