+++ app/templates/errors/404error.html
@@ -6,7 +6,8 @@
 <div class="text-center">
     <h1 class="display-1">Error 404: Page Not Found</h1>
 
-    <p class="">The page you tried to access does not exist. If you believe this is a mistake, please contact the system administrator.</p>
+    <p class="">The page you tried to access does not exist. If you believe this is a mistake,
+                <a href="mailto:heggens@berea.edu,ramsayb2@berea.edu">please contact the system administrators.</a></p>
     <br>
     <a class="btn btn-primary" href="/" role=button>Return Home</a>
 </div>
