+++ app/templates/errors/500error.html
@@ -6,8 +6,8 @@
 {% block app_content %}
 <div class="text-center">
     <h1 class="display-1">Error 500</h1>
-
-    <p class="">Something went wrong on our end, please try again later. If the problem persists, please contact the system administrator.</p>
+    <p class="">Something went wrong on our end, please try again in a few minutes.
+                  If the problem persists, <a href="mailto:heggens@berea.edu,ramsayb2@berea.edu">please contact the system administrators.</a></p>
     <br>
     <a class="btn btn-primary" href="/" role=button>Return Home</a>
 </div>
